+++ app/repositories/managers/__init__.py
@@ -1,0 +1,6 @@
+from .beverage import BeverageManager
+from .ingredient import IngredientManager
+from .order import OrderManager
+from .size import SizeManager
+from .index import IndexManager
+from .base import BaseManager
