+++ app/repositories/managers/size.py
@@ -1,0 +1,8 @@
+from .base import BaseManager
+from ..models import Size
+from ..serializers import SizeSerializer
+
+
+class SizeManager(BaseManager):
+    model = Size
+    serializer = SizeSerializer
