+++ app/repositories/models/order_detail.py
@@ -1,0 +1,11 @@
+from app.plugins import db
+
+
+class OrderDetail(db.Model):
+    _id = db.Column(db.Integer, primary_key=True)
+    price = db.Column(db.Float)
+    order_id = db.Column(db.Integer, db.ForeignKey('order._id'))
+    ingredient_id = db.Column(db.Integer, db.ForeignKey('ingredient._id'))
+    ingredient = db.relationship('Ingredient', backref=db.backref('ingredient'))
+    beverage_id = db.Column(db.Integer, db.ForeignKey('beverage._id'))
+    beverage = db.relationship('Beverage', backref=db.backref('beverage'))
