+++ setup.cfg
@@ -3,4 +3,4 @@ relative_files = True
 
 [flake8]
 max-line-length = 160
-exclude = */services/__init__.py, */controllers/__init__.py
+exclude = */services/__init__.py, */controllers/__init__.py, */repositories/*/__init__.py
